refactor: refusals.tsx is one socket — the sentence, the keying and the row
architecture-first-principles + hickey/lowy over the diff.

FRAGMENTATION (hickey L2), the headline: the module already owned the refusal
SENTENCE and the memo that keeps it from being read out loud twice — and then
each of the three doors wrote out the `<Index>` and the alarmed row itself.
Those are two halves of ONE promise kept in two places: a memo whose value
holds is announced again anyway by a list drawn with `<Key>`, and a door that
remembered the memo could still forget the keying. The module's own header
said as much ("the `<Index>` at each site is the second half of the same
promise") — a rule kept by convention at three sites, which is exactly the
"update X when Y changes" pattern.

So `refusals.ts` becomes `refusals.tsx` and exports ONE thing (lowy §6.5, one
entry point per coherent concept): `<Refusals of={…} class={…} testid={…}/>`.
A door hands over the raw `Refusal`s its own reading answered with; `refusalLines`
is internal. What a door still says is what only it knows — where the rows sit
and what it calls them.

NO-OP, recorded rather than dismissed: the two `ALERT_ROW` constants are not
collapsed into a shared class string. The canonical in-repo pattern is a
caller spelling its own box inline (eleven other `SaidLine` call sites do), and
`SaidLine`'s own header rules that layout is the caller's — a shared band
string would be a parallel receptacle for something already ruled on, and the
two differ in the padding each panel's rows set.
