+++ packages/web/README.md
@@ -44,7 +44,7 @@ The two spellings that are correct, and when each is:
 - **`<Index>`** where the list is POSITIONAL and short — the shortlists (`search/Shortlist.tsx`, `complete/Completions.tsx`, the header box's rows, the composer's `@` rows), and the refusal lines under a query. The element belongs to the slot and its contents are patched, which is what a cursor walking by index already means.
 - **`<For>` over STRINGS**, which key by value: the transcript's rows and the month grid. An unchanged string is the same key, so the list holds still without anything being declared about it.
 
-The same fact reaches derivations, not only lists — a memo that reads a wire array from an UNDECLARED member notifies on every frame whatever the frame said — and the answer there is the one `served.tsx` already gives: an `equals` that compares the value, so the memo's identity means "the answer changed" rather than "a frame arrived". `names.ts` does it for the one table the leaves of a page all read, and its own suite now measures what that is still worth: `names` carries no `key`, so it merges by position and neither a repeated frame nor a row's change wakes it at all — what the comparison still earns is the NAVIGATION, whose first frame has nothing to merge into and wakes every reader whatever it says. `refusals.ts` does it for the refusal SENTENCES, which is a derivation over a client-side parse rather than over a frame, and is what keeps `role="alert"` from reading the same refusal out loud again on the next keystroke. Both hold a plain COPY, because a comparator needs two values and an array off the store is the same array frame after frame.
+The same fact reaches derivations, not only lists — a memo that reads a wire array from an UNDECLARED member notifies on every frame whatever the frame said — and the answer there is the one `served.tsx` already gives: an `equals` that compares the value, so the memo's identity means "the answer changed" rather than "a frame arrived". `names.ts` does it for the one table the leaves of a page all read, and its own suite now measures what that is still worth: `names` carries no `key`, so it merges by position and neither a repeated frame nor a row's change wakes it at all — what the comparison still earns is the NAVIGATION, whose first frame has nothing to merge into and wakes every reader whatever it says. `refusals.tsx` does it for the refusal SENTENCES, which is a derivation over a client-side parse rather than over a frame, and is what keeps `role="alert"` from reading the same refusal out loud again on the next keystroke. Both hold a plain COPY, because a comparator needs two values and an array off the store is the same array frame after frame.
 
 What a `<Key>` still pays per frame is now what the declaration says it pays. `keyArray` sets a row's item signal when that row's object was replaced — which, undeclared, is every row on every frame, so each row's bindings re-ran even though its DOM survived (the audit's 2.11, the one finding with no client-side fix). Declared, the row objects survive: `Tree.browsertest.ts` measures it on both sides, and over three rows a repeated frame went from **3 of 3** bindings re-run to **0**, one row's title from **3** to **1**, and a reorder from **3** to **0**.
 
@@ -479,7 +479,9 @@ The filter rides the ADDRESS (`routes.ts`'s `?q=`), so a narrowed page is a link
 
 The matching these doors get is entirely the server's — the same `Query.search` reading an agent's `search_nodes` gets — so there is deliberately no matcher IN THIS DIRECTORY (`palette/items.ts` says why), and since `search-server-side` there is none anywhere else in this client either: the filter asks the other half of the same reading (`filter/asking.ts`), and the composer's `@` list is a caller of `nodes.ts` like the rest. One reading, six doors — never a second one. WHEN each of those doors asks is one rule too, and one file (`settled.ts`). What the answer carries besides the rows is `refusals`: a known operator with an unknown value, in the grammar's own words, drawn by both doors in a row of its own. It is a DIFFERENT slot from `failure` for the reason `failure` is a different slot from the palette's `>` ask — a refused call and a refused query are two pieces of news, and a reader shown one in the other's sentence has been told something untrue.
 
-**And every one of those slots is ONE component.** Each door had hand-rolled its own `role="alert"` band — five of them between the palette and the header box, against the filter bar's single line through `edit/SaidLine.tsx` — so the alarm colour, the `data-tone` fact and the `role`/`aria-live` pair that decides whether a screen reader is interrupted were spelled once per row. They are that component's now, at all three doors: a refusal is why a list is empty, which is the same mood as why a write did not land, and a surface that had to remember the pair is a surface that can forget it. What stays each door's own is WHERE the line sits — a band across the palette at its own padding, a narrower one over the header's panel, a line under the filter bar — because that is the one thing they genuinely differ about. The live-region behaviour the sentence memo buys (`refusals.ts`) survives the swap and is pinned per door in `an_answer_leaves_the_rows_standing.feature`.
+**And every one of those slots is ONE component.** Each door had hand-rolled its own `role="alert"` band — five of them between the palette and the header box, against the filter bar's single line through `edit/SaidLine.tsx` — so the alarm colour, the `data-tone` fact and the `role`/`aria-live` pair that decides whether a screen reader is interrupted were spelled once per row. They are that component's now: a refusal is why a list is empty, which is the same mood as why a write did not land, and a surface that had to remember the pair is a surface that can forget it.
+
+**The REFUSALS go one step further, and `refusals.tsx` is the whole of them.** That module held the sentence and the memo that keeps it from being read out loud twice, and each of the three doors then wrote out the `<Index>` and the alarmed row itself — which is two halves of one promise kept in two places, since a memo whose value holds is announced again anyway by a list drawn with `<Key>`. It exports one thing now, `<Refusals of={…} class={…} testid={…}/>`: a door hands over the raw `Refusal`s its own reading answered with and gets the rows. What stays the door's is what only it knows — where they sit (a band across the palette at its own padding, a narrower one over the header's panel, a line under the filter bar) and what it calls them, since a `data-testid` is a contract with a package this one does not import. The live-region promise is pinned per door in `an_answer_leaves_the_rows_standing.feature`.
 
 ## The keyboard, in one file
